# Tracer performance optimisations (solution 1)