--- a/eth/tracers/testdata/call_tracer_big_slow.json
+++ b/eth/tracers/testdata/call_tracer_big_slow.json
@@ -1,0 +1,81 @@
+{
+  "genesis": {
+    "difficulty": "50486697699375",
+    "extraData": "0xd783010406844765746887676f312e362e32856c696e7578",
+    "gasLimit": "4788482",
+    "hash": "0xf6bbc5bbe34d5c93fd5b4712cd498d1026b8b0f586efefe7fe30231ed6b8a1a5",
+    "miner": "0xbcdfc35b86bedf72f0cda046a3c16829a2ef41d1",
+    "mixHash": "0xabca93555584c0463ee5c212251dd002bb3a93a157e06614276f93de53d4fdb8",
+    "nonce": "0xa64136fcb9c2d4ca",
+    "number": "1719576",
+    "stateRoot": "0xab5eec2177a92d633e282936af66c46e24cfa8f2fdc2b8155f33885f483d06f3",
+    "timestamp": "1466150166",
+    "totalDifficulty": "28295412423546970038",
+    "alloc": {
+      "0xf8bda96b67036ee48107f2a0695ea673479dda56": {
+        "balance": "0x1529e844f9ecdeec",
+        "nonce": "33",
+        "code": "0x",
+        "storage": {}
+      }
+    },
+    "config": {
+      "networkId": 1,
+      "chainId": 61,
+      "eip2FBlock": 1150000,
+      "eip7FBlock": 1150000,
+      "eip150Block": 2500000,
+      "eip155Block": 3000000,
+      "eip160Block": 3000000,
+      "eip161FBlock": 8772000,
+      "eip170FBlock": 8772000,
+      "eip100FBlock": 8772000,
+      "eip140FBlock": 8772000,
+      "eip198FBlock": 8772000,
+      "eip211FBlock": 8772000,
+      "eip212FBlock": 8772000,
+      "eip213FBlock": 8772000,
+      "eip214FBlock": 8772000,
+      "eip658FBlock": 8772000,
+      "eip145FBlock": 9573000,
+      "eip1014FBlock": 9573000,
+      "eip1052FBlock": 9573000,
+      "eip152FBlock": 10500839,
+      "eip1108FBlock": 10500839,
+      "eip1344FBlock": 10500839,
+      "eip1884FBlock": 10500839,
+      "eip2028FBlock": 10500839,
+      "eip2200FBlock": 10500839,
+      "ecip1010PauseBlock": 3000000,
+      "ecip1010Length": 2000000,
+      "ecip1017FBlock": 5000000,
+      "ecip1017EraRounds": 5000000,
+      "ecip1099FBlock": 11700000,
+      "ecbp1100FBlock": 11380000,
+      "disposalBlock": 5900000,
+      "ethash": {},
+      "requireBlockHashes": {
+          "1920000": "0x94365e3a8c0b35089c1d1195081fe7489b528a84b22199c916180db8b28ade7f",
+          "2500000": "0xca12c63534f565899681965528d536c52cb05b7c48e269c2a6cb77ad864d878a"
+        }
+    }
+  },
+  "context": {
+    "number": "1719577",
+    "difficulty": "50486697732143",
+    "timestamp": "1466150178",
+    "gasLimit": "4788484",
+    "miner": "0x2a65aca4d5fc5b5c859090a6c34d164135398226"
+  },
+  "input": "0xf874218504a817c800832318608080a35b620186a05a131560135760016020526000565b600080601f600039601f565b6000f31ba0575fa000a1f06659a7b6d3c7877601519a4997f04293f0dfa0eee6d8cd840c77a04c52ce50719ee2ff7a0c5753f4ee69c0340666f582dbb5148845a354ca726e4a",
+  "result": {
+    "type": "CREATE",
+    "from": "0xf8bda96b67036ee48107f2a0695ea673479dda56",
+    "to": "0xb2e6a2546c45889427757171ab05b8b438525b42",
+    "value": "0x0",
+    "gas": "0x22410c",
+    "gasUsed": "0x20baae",
+    "input": "0x5b620186a05a131560135760016020526000565b600080601f600039601f565b6000f3",
+    "output": "0x"
+  }
+}