… eth/tracers/internal/tracers/assets.go
… eth/tracers/testdata/parity_call_tracer_big_slow.json
… eth/tracers/testdata/parity_call_tracer_callcode_precompiled_fail_hide.json
… eth/tracers/testdata/parity_call_tracer_callcode_precompiled_oog.json
… eth/tracers/testdata/parity_call_tracer_callcode_precompiled_throw.json
… eth/tracers/testdata/parity_call_tracer_create.json
… eth/tracers/testdata/parity_call_tracer_deep_calls.json
… eth/tracers/testdata/parity_call_tracer_delegatecall.json
… eth/tracers/testdata/parity_call_tracer_gas.json
… eth/tracers/testdata/parity_call_tracer_inner_create_oog_outer_throw.json
… eth/tracers/testdata/parity_call_tracer_inner_instafail.json
… eth/tracers/testdata/parity_call_tracer_inner_throw_outer_revert.json
… eth/tracers/testdata/parity_call_tracer_nested_create.json
… eth/tracers/testdata/parity_call_tracer_nested_create2_action_gas.json
… eth/tracers/testdata/parity_call_tracer_nested_create_action_gas.json
… eth/tracers/testdata/parity_call_tracer_nested_create_inerror.json
… eth/tracers/testdata/parity_call_tracer_oog.json
… eth/tracers/testdata/parity_call_tracer_result_output.json
… eth/tracers/testdata/parity_call_tracer_revert.json
… eth/tracers/testdata/parity_call_tracer_revert_reason.json
… eth/tracers/testdata/parity_call_tracer_selfdestruct.json
… eth/tracers/testdata/parity_call_tracer_simple.json
… eth/tracers/testdata/parity_call_tracer_skip_no_balance_error.json
… eth/tracers/testdata/parity_call_tracer_staticcall_precompiled.json
… eth/tracers/testdata/parity_call_tracer_throw.json