This resolves #254.

This PR handles tracer performance optimisations and it is one of the two suggested solutions which has:

- cleaner logic in the `tracer.go` file, as it has no custom logic there, based on a JS tracer
- slower performance than solution 2, as it still has to call a JS method using duktape

Below you can find some sample measurements for each test case, with the important one being the `bigSlow`:

### Test CallTracer

| Test | not optimised | optimised |
|------|-------------:|-----------:|
| bigSlow | 5.45653184s | 3.339744961s |
| create | 2.575857ms | 1.721851ms |
| deepCalls | 69.298277ms | 53.81494ms |
| delegatecall | 13.072359ms | 12.216448ms |
| innerCreateOogOuterThrow | 4.356151ms | 5.754842ms |
| innerInstafail | 1.789353ms | 3.336143ms |
| innerThrowOuterRevert | 15.972036ms | 13.377374ms |
| oog | 5.654414ms | 6.051736ms |
| revert | 2.175ms | 1.472202ms |
| revertReason | 4.936105ms | 4.174434ms |
| simple | 5.427691ms | 3.436775ms |
| throw | 1.909941ms | 3.426322ms |

### Test CallTracerParity

| Test | not optimised | optimised |
|------|-------------:|-----------:|
| bigSlow | 5.377376866s | 4.381227684s |
| create | 2.487462ms | 1.834985ms |
| deepCalls | 80.946559ms | 66.649341ms |
| delegatecall | 14.748616ms | 17.8709ms |
| innerCreateOogOuterThrow | 10.448277ms | 9.212009ms |
| innerInstafail | 2.620816ms | 2.585761ms |
| innerThrowOuterRevert | 15.346743ms | 14.264673ms |
| oog | 5.349929ms | 2.512547ms |
| revertReason | 4.295931ms | 728.27µs |
| revert | 1.33ms | 3.595533ms |
| simple | 4.94809ms | 7.79357ms |
| throw | 3.851077ms | 5.416296ms |