@ziogaschr 2020-12-21
Closing this one as #256 has been preferred.